Worked on index.html (hero,nav)
HTML
I worked on the nav and hero section. id tags and buttons

 I started to link the CSS to this section. I also started the blocks that style the page in CSS. I also started the button styling

I need help on moving my nav to the right side of the page as well as having them spaced like they are in figma. I also need to see if my spam tags were correct or not when trying to create color blocks for style. I also need help with the code for my buttons. They are rectangle with sharp edges but i want a rounded edge. I have the code there for it but it is not actively working. I am free tomorrow any time after 3pm to discuss these changes. Also you can just slack me and I can clarify anything tat you are unsure of.