Media Queries attempt/ HTML
I got a monitor so I could see the code better. I worked on trying to do the media Queries. I really struggles because it seemed as if the changes I made in the code did not come up on dev tools even after I reloaded and reopened the HTML. 

What I did 

HTML
-got rid of the extra email newsletter on the contact page
-double checked my favicon because it wasn't coming up (it is now) 

CSS
-worked on the media Queries (I really struggled because I couldn't actively see my changes in Dev tools)
-I reworked the Tour page right panel to make sure it fit a desktop screen
- I reworked the bio page (still needs some work) 
- I reworked some of the rectangles in in the contact page so everything should be set for the desktop size.

What I need help on
- Rounding out my photo and Text box in the bio.html
-Mainly I need help in Media Queries Everything I seemed to do was not correct. I watched your video as well as sought help from W3schools and other websites.